requirements.txt
numpy
requests
sounddevice==0.4.6
openai-whisper
urllib3<2